backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
supabase==2.0.2
asyncpg==0.29.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# AI Provider integrations (direct HTTP calls)
openai==1.6.1
anthropic==0.8.1

# Caching and rate limiting
redis==5.0.1
slowapi==0.1.9

# Configuration and environment
pydantic-settings==2.0.3
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Monitoring and logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
